VendorManagementSystem/requirement.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
click==8.1.7
cssselect==1.2.0
Django==5.0.4
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.7
lxml==5.2.1
Markdown==3.6
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
Pygments==2.18.0
pyquery==2.0.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
yarl==1.9.4
